refactor[next]: split ffront test_execution.py into feature files
Carve the 1469-line catch-all
tests/next_tests/integration_tests/feature_tests/ffront_tests/test_execution.py
into per-feature files, a pure relocation (collection count unchanged at 7789):

- new: test_scan.py, test_tuples.py, test_astype.py, test_domain.py,
  test_scalar_args.py, test_cartesian_shifts.py, test_closure_vars.py
- fold unstructured shift/neighbor-reduction tests into test_gt4py_builtins.py
- rename test_scalar_if.py -> test_conditionals.py and add the ternary tests
- shrink test_execution.py to basics and rename it test_basic.py

No test bodies or markers changed; every moved test keeps its uses_* markers.
